Extract devo components, add trad devo sharing
