+++ src/screens/HomeScreen.tsx
@@ -1,31 +1,10 @@
 import { Ionicons } from "@expo/vector-icons";
-import React, { useEffect, useRef, useState } from "react";
-import {
-  ActivityIndicator,
-  FlatList,
-  Keyboard,
-  KeyboardAvoidingView,
-  Platform,
-  RefreshControl,
-  ScrollView,
-  StyleSheet,
-  Text,
-  TextInput,
-  TouchableOpacity,
-  View,
-} from "react-native";
-import { Snackbar } from "react-native-paper";
-import apiConfig from "../../apiConfig";
-import { auth } from "../../firebase";
-import { TTradDevo } from "../../types";
+import React, { useState } from "react";
+import { StyleSheet, Text, TouchableOpacity, View } from "react-native";
 import { Container } from "../components/Container";
-import VerseContainer from "../components/VerseContainer";
-import { useApi } from "../hooks/useApi";
-import { useRequestReview } from "../hooks/useRequestReview";
-import { useTradDevos } from "../hooks/useTradDevos";
-import useStore from "../Store";
+import { PersonalizedDevotional } from "../components/PersonalizedDevotional";
+import { TraditionalDevotionals } from "../components/TraditionalDevotionals";
 import colors from "../styles/colors";
-import { formatDate } from "../utils";
 
 const HomeScreen: React.FC = () => {
   const [devoType, setDevoType] = useState<"traditional" | "personalized">(
@@ -87,232 +66,11 @@ const HomeScreen: React.FC = () => {
       {devoType === "personalized" ? (
         <PersonalizedDevotional />
       ) : (
-        <TraditionalDevotional />
+        <TraditionalDevotionals />
       )}
     </Container>
   );
 };
-
-function PersonalizedDevotional() {
-  const {
-    promptStart,
-    input,
-    setInput,
-    setDevotional,
-    setPromptId,
-    error,
-    setError,
-  } = useStore();
-  const inputRef = useRef<TextInput>(null);
-  const {
-    isLoading,
-    data,
-    fetchData: getDevotional,
-  } = useApi<{
-    response: string;
-    promptId: string;
-  }>(`${apiConfig.apiUrl}/getGpt3Response`, {
-    method: "POST",
-    body: JSON.stringify({
-      userId: auth.currentUser?.uid,
-      prompt: input,
-    }),
-    headers: { "Content-Type": "application/json" },
-  });
-  const { requestReview } = useRequestReview();
-
-  useEffect(() => {
-    let requestReviewTimeout: any;
-
-    if (data && data.response && data.promptId) {
-      setPromptId(data.promptId);
-      setDevotional(data.response);
-      requestReviewTimeout = setTimeout(() => {
-        requestReview();
-      }, 10000);
-    }
-
-    // Call requestReview ten seconds after data comes back
-    return () => {
-      clearTimeout(requestReviewTimeout);
-    };
-  }, [data]);
-
-  useEffect(() => {
-    if (promptStart) {
-      let formattedPrompt = promptStart.split(".")[0] + " ";
-      setInput(formattedPrompt);
-      if (inputRef.current) inputRef.current.focus();
-    } else {
-      setInput("");
-    }
-  }, [promptStart]);
-
-  const submit = () => {
-    getDevotional();
-    setDevotional("");
-    Keyboard.dismiss();
-  };
-
-  // TODO: Refactor to use Container, delegate Snackbar error handling to Container
-  return (
-    <KeyboardAvoidingView
-      behavior={Platform.OS === "ios" ? "padding" : "height"}
-      style={{ flex: 1 }}
-    >
-      <ScrollView
-        style={styles.scroller}
-        contentContainerStyle={{ flexGrow: 1 }}
-        keyboardShouldPersistTaps="handled"
-      >
-        <View style={styles.container}>
-          <View style={{ marginBottom: 20, alignItems: "center" }}>
-            <Text
-              style={{
-                color: colors.black,
-                fontSize: 18,
-                fontWeight: "500",
-                paddingVertical: 10,
-              }}
-            >
-              Ask a question
-            </Text>
-            <Text style={{ color: colors.black, fontSize: 16 }}>
-              Receive a tailored devotional just for you
-            </Text>
-          </View>
-          <TextInput
-            ref={inputRef}
-            style={[styles.input]}
-            placeholder="Type your question or describe your situation here..."
-            placeholderTextColor="#999"
-            onChangeText={(text) => setInput(text)}
-            value={input}
-            multiline
-          />
-          <View style={styles.buttonRow}>
-            <TouchableOpacity
-              onPress={submit}
-              style={[
-                styles.button,
-                { opacity: input && !isLoading ? 1 : 0.4 },
-              ]}
-              disabled={!input || isLoading}
-            >
-              <Text style={styles.buttonText}>Get Guidance</Text>
-            </TouchableOpacity>
-          </View>
-          <VerseContainer isLoading={isLoading} />
-        </View>
-      </ScrollView>
-      <Snackbar
-        visible={!!error}
-        onDismiss={() => setError(null)}
-        action={{
-          label: "Dismiss",
-          onPress: () => setError(null),
-        }}
-      >
-        {error}
-      </Snackbar>
-    </KeyboardAvoidingView>
-  );
-}
-
-function TraditionalDevotional() {
-  const { tradDevos, loading, refreshing, setRefreshing } = useTradDevos();
-
-  if (loading) {
-    return (
-      <View style={{ flex: 1, justifyContent: "center", alignItems: "center" }}>
-        <ActivityIndicator />
-      </View>
-    );
-  }
-
-  return (
-    <View>
-      <FlatList
-        data={tradDevos}
-        renderItem={({ item }) => <DevotionalCard devotional={item} />}
-        keyExtractor={(item) => item.id}
-        style={{ width: "100%", paddingHorizontal: 20 }}
-        ListEmptyComponent={() => (
-          <View style={{ alignItems: "center", marginTop: 24 }}>
-            <Text style={{ fontSize: 18, color: "#999" }}>
-              No devotions found
-            </Text>
-          </View>
-        )}
-        ListFooterComponent={() => <View style={{ height: 48 }} />}
-        refreshControl={
-          <RefreshControl
-            refreshing={refreshing}
-            onRefresh={() => setRefreshing(true)}
-          />
-        }
-      />
-    </View>
-  );
-}
-
-function DevotionalCard({ devotional }: { devotional: TTradDevo }) {
-  const [isExpanded, setIsExpanded] = useState(false);
-
-  const handleToggleExpanded = () => {
-    setIsExpanded(!isExpanded);
-  };
-
-  return (
-    <View
-      style={{
-        borderRadius: 12,
-        paddingVertical: 24,
-        paddingHorizontal: 4,
-        borderBottomColor: colors.lightBlue,
-        borderBottomWidth: 2,
-      }}
-    >
-      <TouchableOpacity onPress={handleToggleExpanded}>
-        <Text style={{ fontSize: 18, fontWeight: "bold", marginBottom: 4 }}>
-          {devotional.title}
-        </Text>
-        <Text style={{ fontSize: 14, color: "#999", marginBottom: 12 }}>
-          {formatDate(devotional.createdAt)}
-        </Text>
-        <Text
-          style={{
-            fontSize: 16,
-            lineHeight: 24,
-            marginBottom: isExpanded ? 12 : 0,
-            fontStyle: "italic",
-          }}
-        >
-          {devotional.input}
-        </Text>
-      </TouchableOpacity>
-      {isExpanded && (
-        <>
-          <View
-            style={{
-              borderTopColor: "#eee",
-              borderTopWidth: 1,
-              paddingTop: 12,
-              marginBottom: 12,
-              // TODO: Make section more visually distinct
-              // TODO: Support verse highlighting and clickthrough actions here
-              // TODO: Only have a single devo expanded at a time
-            }}
-          >
-            <Text style={{ fontSize: 16, lineHeight: 24 }}>
-              {devotional.response}
-            </Text>
-          </View>
-        </>
-      )}
-    </View>
-  );
-}
 
 const styles = StyleSheet.create({
   scroller: {
